Create PlayerModel (#11)
